Move monkey patches to own directory
