+++ lib/sequent/core/helpers/attribute_support.rb
@@ -1,47 +1,5 @@
 require 'active_support'
-# TODO: Move this into a separate core_ext folder like for instance Rails does.
-# WARNING: Monkey patches below...
-class Symbol
-  def self.deserialize_from_json(value)
-    value.try(:to_sym)
-  end
-end
-
-class String
-  def self.deserialize_from_json(value)
-    value
-  end
-end
-
-class Integer
-  def self.deserialize_from_json(value)
-    value.blank? ? nil : value.to_i
-  end
-end
-
-class Boolean
-  def self.deserialize_from_json(value)
-    value.nil? ? nil : (value.present? ? value : false)
-  end
-end
-
-class Date
-  def self.deserialize_from_json(value)
-    value.nil? ? nil : Date.iso8601(value.dup)
-  end
-end
-
-class DateTime
-  def self.deserialize_from_json(value)
-    value.nil? ? nil : DateTime.iso8601(value.dup)
-  end
-end
-
-class Array
-  def self.deserialize_from_json(value)
-    value
-  end
-end
+require_relative '../ext/ext'
 
 module Sequent
   module Core
